Move schema.js to src folder
